// api/_lib/types.ts
export type FileType = "png" | "jpeg";
export type Theme = "light" | "dark";

export interface ParsedRequest {
  fileType: FileType;
  side: string;
  market: string;
  pnl: string;
  avgEntry: string;
  markPrice: string;
  theme: Theme;
  md: boolean;
  fontSize: string;
  images: string[];
  widths: string[];
  heights: string[];
}